+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/facade/DBVariantContextAnnotatorFactory.java
@@ -87,4 +87,23 @@ public class DBVariantContextAnnotatorFactory {
 		return new DBVariantContextAnnotator(new ClinVarAnnotationDriver(vcfDBPath, fastaRefPath, options), options);
 	}
 
+
+	/**
+	 * Construct Cosmic {@link VariantContext} annotator factory.
+	 *
+	 * @param vcfDBPath
+	 *            Path to gzip-compressed, normalized and tbi-indexed Cosmic file to use for the annotation
+	 * @param fastaRefPath
+	 *            Path to reference FASTA file
+	 * @param options
+	 *            Configuration for the variant context annotation
+	 * @return Preconfigured {@link DBVariantContextAnnotator} object
+	 * @throws JannovarVarDBException
+	 *             on problems loading the resources
+	 */
+	public DBVariantContextAnnotator constructCosmic(String vcfDBPath, String fastaRefPath,
+			DBAnnotationOptions options) throws JannovarVarDBException {
+		return new DBVariantContextAnnotator(new CosmicAnnotationDriver(vcfDBPath, fastaRefPath, options), options);
+	}
+
 }