+++ jannovar-vardbs/src/test/java/de/charite/compbio/jannovar/vardbs/dbsnp/DBSNPAnnotationDriverReportAlsoOverlappingTest.java
@@ -44,9 +44,9 @@ public class DBSNPAnnotationDriverReportAlsoOverlappingTest extends DBSNPAnnotat
 
 		// Check header after extension
 		Assert.assertEquals(0, header.getFilterLines().size());
-		Assert.assertEquals(10, header.getInfoHeaderLines().size());
+		Assert.assertEquals(12, header.getInfoHeaderLines().size());
 		Assert.assertEquals(0, header.getFormatHeaderLines().size());
-		Assert.assertEquals(10, header.getIDHeaderLines().size());
+		Assert.assertEquals(12, header.getIDHeaderLines().size());
 		Assert.assertEquals(0, header.getOtherHeaderLines().size());
 
 		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_COMMON"));
@@ -54,11 +54,13 @@ public class DBSNPAnnotationDriverReportAlsoOverlappingTest extends DBSNPAnnotat
 		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_G5"));
 		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_G5A"));
 		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_IDS"));
+		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_SAO"));
 		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_OVL_COMMON"));
 		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_OVL_CAF"));
 		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_OVL_G5"));
 		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_OVL_G5A"));
 		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_OVL_IDS"));
+		Assert.assertNotNull(header.getInfoHeaderLine("DBSNP_OVL_SAO"));
 	}
 
 	@Test